audit(fragment): fix segment_num 8-bit-vs-12-bit width bug, close Group K (issue #256) (#283)
Full-catalog audit follow-up, batch 27 (Group K, misc triage):

1. Real bug fix: RCP_FRAGMENT_MAX_INTERMEDIATE_SEGMENTS/segment_num
   was documented and stored as an 8-bit field (256 values), but
   acf.h's own bit layout (read_size_or_segment_num[11:8] in byte 6,
   [7:0] in byte 7) is unambiguously 12 bits (0..4095), independently
   confirmed by acf.c's own already-tested bit-masking constants.
   segment_num was uint8_t throughout fragment.h/fragment.c, silently
   truncating any real decoded value above 255 -- a genuine
   wire-compatibility bug. Widened segment_num/expected_segment_num to
   uint16_t, corrected the macro to 4096u, updated all doc comments.
   New test proves values above 255 round-trip correctly; mutation-
   tested twice (field-width revert -> clean assertion failure;
   bounds-check loosening -> SIGSEGV from a real buffer overrun) to
   confirm both the fix and its guard are load-bearing.

2. REQ-SCHED-001..008 verified as a false positive: already scope
   "tc18" with accurate citations since v0.69.0, no fix needed.

Closes Group K, and with it Groups H-K of issue #256's 156-finding
catalog in full. 65/65 both trees. cfusa check: 0 errors. cfusa
trace: 0/1024 untested, 100% req-coverage, 100% sec-tested.

+++ src/fragment.c
@@ -66,7 +66,7 @@ rcp_fragment_errc_t rcp_fragment_plan(size_t payload_len, size_t max_fragment_pa
         out_segments[i].offset      = off;
         out_segments[i].len         = max_fragment_payload;
         out_segments[i].ms          = true;
-        out_segments[i].segment_num = (uint8_t)i;
+        out_segments[i].segment_num = (uint16_t)i;
         off += max_fragment_payload;
     }
 
@@ -153,7 +153,7 @@ static bool append(rcp_fragment_reassembler_t *r, const uint8_t *payload, size_t
 //cfusa:req REQ-FRAG-015
 //cfusa:req REQ-FRAG-016
 rcp_fragment_reasm_result_t rcp_fragment_reassembler_feed(rcp_fragment_reassembler_t *r,
-                                                            bool ms, uint8_t segment_num,
+                                                            bool ms, uint16_t segment_num,
                                                             const uint8_t *payload,
                                                             size_t payload_len)
 {
@@ -180,7 +180,7 @@ rcp_fragment_reasm_result_t rcp_fragment_reassembler_feed(rcp_fragment_reassembl
     if (!append(r, payload, payload_len)) return RCP_FRAGMENT_REASM_ERR_ALLOC;
 
     if (ms) {
-        r->expected_segment_num = (uint8_t)(r->expected_segment_num + 1);
+        r->expected_segment_num = (uint16_t)(r->expected_segment_num + 1);
         return RCP_FRAGMENT_REASM_CONTINUE;
     }
